image_captioning/requirements.txt
tensorflow==2.18.0
numpy==2.0.2
pillow==11.0.0
flask==3.0.3
gunicorn==23.0.0
transformers==4.46.2
tf-keras==2.18.0